+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_env_telemetry.py
@@ -43,7 +43,7 @@ class ContainerappEnvTelemetryScenarioTest(ScenarioTest):
             JMESPathCheck('[0].name', env_name),
         ])
 
-        self.cmd(f'containerapp env telemetry set -g {resource_group} -n {env_name} --app-insights-connection-string {ai_conn_str} --open-telemetry-dataDog-site {data_dog_site} --open-telemetry-dataDog-key {data_dog_key} --open-telemetry-traces-destinations {traces_destinations} --open-telemetry-logs-destinations {logs_destinations} --open-telemetry-metrics-destinations {metrics_destinations}')
+        self.cmd(f'containerapp env telemetry set -g {resource_group} -n {env_name} --app-insights-connection-string {ai_conn_str} --open-telemetry-data-dog-site {data_dog_site} --open-telemetry-data-dog-key {data_dog_key} --open-telemetry-traces-destinations {traces_destinations} --open-telemetry-logs-destinations {logs_destinations} --open-telemetry-metrics-destinations {metrics_destinations}')
         containerapp_env = self.cmd('containerapp env show -g {} -n {}'.format(resource_group, env_name)).get_output_in_json()
 
         while containerapp_env["properties"]["provisioningState"].lower() == "waiting":